Add checking allowanonymous attribute. Fixed problem with data types on model.

+++ BudgetManager.Domain/DTOs/Queries/StockTradesGroupedMonth.cs
@@ -5,8 +5,8 @@
         public int StockTickerId { get; set; }
         public int TradeYear { get; set; }
         public int TradeMonth { get; set; }
-        public int TradeSize { get; set; }
-        public decimal TradeValue { get; set; }
-        public int AccumulatedTradeSize { get; set; }
+        public double TradeSize { get; set; }
+        public double TradeValue { get; set; }
+        public double AccumulatedTradeSize { get; set; }
     }
 }
